menu: actually update the curses menu
This is gonna be slow for bigger lists, since we reconstruct a new MENU
with every update.

Next step here is to make this work using set_menu_items(), which only
segfaulted on me (list isn't updates and segfault after 7 items. Odd).

The step after that is to make us not have to reconstruct the list of
items every update. We just want to append another ITEM* to the menu.