menu: finish the first draft of the menu header
The "header" here being the top row of the table-like structure the menu
paints, so that the user knows which column is the title, authors, etc.

Now we need to use these widths when printing the items;
we don't want the text to escape outside of it's assigned area.
If the text doesn't fit, we can just use '~' as the last character ala
ranger.