menu: print out the whole menu (6 cols for now)
print_header() has been improved, but it's still not utilizing _all_
width possible. I'm not sure how to do that yet. This'll work in the
meantime.

The menu now also prints out six columns of information per item. But
instead of printing item-by-item as we did before, we now print
column-by-column, printing each related string under each.

While this might be seen as an odd change, it simplified the alignment
of the strings under the header by having related pieces of code bunched
close together.

This is certainly not the final version of the menu, but it's somewhere
to start from.