screens: only allow painting on cells that you own
The thing that make this take so long was that the footer was being
printed by the multiselect_menu. So next up, we'll let the screen_butler
manage it.

Since all termbox interaction is now done in the base screen, we should remake
base::change_cell to not take termbox-related arguments. But this isn't
critical right now.

Resolves #11