screen/menu: more work and investigation regarding #12
Alright, multiselect_menu is pretty shit. Not only does it _not_ use the
full width to our advantage, it sometimes writes outside of it's allotted
cells (a heinous crime).

First in order should be to rewrite multiselect_menu to remedy this, but
I've gotten quite burned out on TUI code for the time being. Future me
will take care of it, I'm sure.

I also cannot help the feeling that termbox is acting up: if I resize
the window the last column will some times be written out even though
it obviously wont fit. But that's probably just the can-we-fit-it-check
acting up.