butler/script: correctly catch py::error_already_set
This requires us to change a few things. Rather than letter the logger
just print on top of the TUI, we instead store the log messages in a
buffer that is empied (printed out) upon logger destruction.

Effectively, this means that any logs done when the TUI is up are
instead printed during program termination, when the TUI is down.