abstract termbox colours/attributes into own header
Now, I'd rather overload

    inline colour operator=(attribute attr)
    {
        return colour::white | attr;
    }

But since enums aren't classes, this seems to require to wrap everything
in a class. Perhaps that would be a better idea (esp. as we might
upgrade to 256 colours post-v1.0.0), but for now we'll resort to
function overloads in screen::base and butler::screen_butler.

This commit also makes the rules described in termbox.h strongly typed:

    /* Attributes, it is possible to use multiple attributes by combining them
     * using bitwise OR ('|'). Although, colors cannot be combined. But you can
     * combine attributes and a single color. See also struct tb_cell's fg and bg
     * fields.
     */