let screen::base handle all padding
screen::multiselect_menu still relies on the padding variables to get
the right alignment. We'll have to review that. Making the padding
variables private in screen::base would be for the best, but then we
might need to write a function that gives us the right-most/down-most
column/line we can write on.

Resolves #15