logger: log all entries to screen if able
To do this, we create a buffer that stores whatever entries are created
before the TUI is up and running. During screen construction, when we

    sink_->set_screen_butler(butler);

we immidiately flush the buffer's content to the log screen. If this
isn't possible (we're unable to create a screen for whatever reason) the
entries are instead flushed to stdout or stdcerr during logger
destruction.