butler/screen: notify user about unread logs
This prints a small right-aligned string in the context-sensitive info
bar about unread messages. The background colour of the message matches
that of the worst yet unread message — e.g. if one of the unread
messages is an error, the notice will be printed in red.

This commit also introduces more useless updating. Since we don't want
to wait for an action to notify the user about any possible unread logs,
we have to update the screens even when we just add an entry to the
buffer.