check termination status before HTTP requests
Before every HTTP request, check if bookwyrm is terminating. If she is,
raise GeneratorExit; the script butler considers this a normal script
termination.
    This allows us to exit bookwyrm before all queries have been made,
but it's too slow, since we may be inside a request when wanting to
terminate. We could potentially use eventlet.Timeout(n), but requires
boiler-plate, which I want to avoid at all costs.
    We should also probably use a custom exception here, but according
to pybind11's documentation, when calling Python from C++ the only
exception we can get is a pybin11::error_already_set, which contain a
Python exception.