Refactor backend into its own library
, among other things. While in the process, I cleaned CMake build file
up, improved the names of many variables, and a ton more which I cannot
recall. I should probably have split this into multiple commits, but the
bulk of the work was done yesterday, during which I couldn't be arsed.
    Fortunately, all bugs found during refactor have also been fixed. One
also appeared: some times when terminating the program, Python will
throw a std::run_time error. This isn't really a problem, but since I
haven't found a way to catch it, it doesn't look very good. This is a
result of detaching the threads in the backend upon termination.

Resolves #14