switch to MIT license
I realised yesternight I don't care what people do with my software.
I'd release bookwyrm in the public domain via Unlicense, but I'm not
sure that is possible since it depends on software not in the public
domain (namely BSD-2, BSD-3, MIT and GPL2 — the latter being my own
repo, but a fork of GPL2 software).