drop spdlog dep. and reimplement logger
In a broken state mind, but the whole code-base is subject to another
modulization; the TUI too shall be be built seperately. Directory
structure will change also, where src/ and include/ shall be merged.
While doing this, the logger shall be fixed but I have a sneaking
suspicion that it may not even be needed.