tui: fix logger behavior and merge into tui.c{pp,hh}
The plan was to greatly simplify it, but I wasn't able to do this.
There is enough code to keep it outside of the main tui class, but it
also needs access to its instance. The logger has seen a modicum of
improvements at least. Instead of a weak_ptr to the tui itself, that
weak_ptr now hold the log screen instead, and a std::function predicate
is used to check if the log screen is focused or not.
    I was glad halfway through this when I ditched the weak_ptr, but it
is ultimately needed because the logger's destructor must run when
NCurses has been terminated.
    Some smaller issues still remain, but that's for another commit.

Resolves #44