store found items in an std::unordered_set instead
With an unordered_set set, the index menu has the interesting effect of
inserting newly found items at different parts of the list (in each
underlying bucket, I presume). This was not intended. While looking
rather cool, it will be annoying for a user. The std::unordered_set will
be replaced with an std::set in a future commit.

screen::multiselect_menu's usage of an std::set<int> of wanted items must
be revisited. A lot of std::next(items_.cbegin(), idx) was used to
replace items_[idx], which is a bad way of doing things. We should hold
iterators to the wanted items in the std::set of items instead.

Related to #30