display how many plugins are still running
Related to #53

A quick-and-dirty implementation. In the future, it might be better to
have a two-way communication between the front and backend, instead of
constructing the TUI with a vanity argument.

The backend does not update the frontend upon the termination of a
plugin; the user must manually update the TUI for the plugin counter to
reach zero.