implement a test framework
Plugins can now be added under test/core/ (more categories to be
implemented sooner) where a simple wrapper around the plugin_handler
class will load and run this module. A passing regexes will then be
matched with the test output.

As of this commit, we only test if exceptions are properly logged. More
to come.

Testing framework structure credit to [0].

[0] <https://gitlab.cern.ch/allpix-squared/allpix-squared>